generate lists of media files to download, per project
Two lists are gnerated, one of files uploaded to the project repo,
and one of files that should be in the foreign repo but are used on
the project. Note that this secon list may include files that don't
exist (see notes.txt).

No deletion prep files are produced yet, in spite of comments in the
methods saying so. That will be done when deletion is implemented.