add stubs for methods to gennerate lists of downloads/deletes in the case where we have a previous run of this script and can skip doing a huge walk through the local media directory tree.
The commit message on the previous commit was so crappy I'll add
more here: in the previous commit due to pylint whines, some methods
were moved out of the main script into the Sync module into a renamed
Projects class where they belong anyways. The list of projects to
process is now handled there, for less repetition. The rest was
finding ways to pare down the list of local variables, args passed, etc
to various methods.