draft adding --full option, and using incremental delete/get files
None of this is tested, and it all makes pylint sad too, but parking it
here so it doesn't get lost on the laptop.

The idea is that --full means we generate full local media lists,
full delete and retrieve lists, even if there have been previous
runs for a wiki.
Without --full, we look for a previous run, and if we find it the
we use the incremental delete and retrieve lists.
Specifying --full excludes use of --continue, and vice versa.